Refactor snake canvas
