+++ components/Snake/SnakeGame.tsx
@@ -1,18 +1,11 @@
 "use client";
 import "@/app/globals.css";
 import React, { useRef, useEffect, useState, useCallback } from "react";
-import {
-  calculateTotalScore,
-  generateFoodPosition,
-  parseScreensConfig,
-  isOppositeDirection,
-} from "@/lib/snake/game";
+import { calculateTotalScore, generateFoodPosition, isOppositeDirection } from "@/lib/snake/game";
 import Gesture from "@/public/gesture.svg";
-import resolveConfig from "tailwindcss/resolveConfig";
 import { orbitron } from "@/lib/fonts";
-import tailwindConfig from "../../tailwind.config";
+import { useCanvasSetup } from "@/lib/snake/useCanvasSetup";
 
-const fullConfig = resolveConfig(tailwindConfig);
 const SNAKE_COLOR = "#3acfd5";
 const FOOD_COLOR = "#ffb3b3";
 const LEVEL_SPEED = 20;
@@ -29,16 +22,6 @@ interface SnakeProps {
 
 export default function SnakeGame({ win, gameOver, level, score, highScore }: SnakeProps) {
   const canvasRef = useRef<HTMLCanvasElement | null>(null);
-  const canvasConfigPrevRef = useRef<{
-    width: number;
-    height: number;
-    unitSize: number;
-  } | null>(null);
-  const canvasConfigRef = useRef<{
-    width: number;
-    height: number;
-    unitSize: number;
-  } | null>(null);
   //todo join to one
   const [startX, setStartX] = useState<number | null>(null);
   const [startY, setStartY] = useState<number | null>(null);
@@ -51,6 +34,7 @@ export default function SnakeGame({ win, gameOver, level, score, highScore }: Sn
   }>({ x: 0, y: 0 });
   const snakeRef = useRef(snake);
   const foodRef = useRef(food);
+  const { canvasConfigRef } = useCanvasSetup(canvasRef, snakeRef, foodRef, setSnake, setFood);
 
   const [direction, setDirection] = useState<{ x: number; y: number }>({
     x: 0,
@@ -67,95 +51,6 @@ export default function SnakeGame({ win, gameOver, level, score, highScore }: Sn
   useEffect(() => {
     foodRef.current = food;
   }, [food]);
-
-  const setupCanvas = useCallback(() => {
-    let unitSize;
-
-    const UNITS_WIDE = 31;
-    const UNITS_TALL = 24;
-
-    if (window.innerWidth >= parseScreensConfig(fullConfig.theme.screens.xl)) {
-      // Desktop
-      unitSize = 30;
-    } else if (window.innerWidth >= parseScreensConfig(fullConfig.theme.screens.md)) {
-      // Tablet
-      unitSize = 22;
-    } else {
-      // Mobile
-      unitSize = 12;
-    }
-
-    if (canvasConfigPrevRef.current?.unitSize === unitSize && snakeRef.current) {
-      return;
-    }
-
-    const width = unitSize * UNITS_WIDE;
-    const height = unitSize * UNITS_TALL;
-    canvasConfigPrevRef.current = { height: 0, unitSize: 0, width: 0, ...canvasConfigRef.current };
-    canvasConfigRef.current = { width, height, unitSize };
-
-    const defaultSnakePosition = 8;
-
-    if (!snakeRef.current) {
-      setSnake([
-        {
-          x: defaultSnakePosition * unitSize,
-          y: defaultSnakePosition * unitSize,
-        },
-        {
-          x: defaultSnakePosition * unitSize - unitSize,
-          y: defaultSnakePosition * unitSize,
-        },
-        {
-          x: defaultSnakePosition * unitSize - 2 * unitSize,
-          y: defaultSnakePosition * unitSize,
-        },
-      ]);
-      setFood(generateFoodPosition(width, height, unitSize, snakeRef.current ?? []));
-    } else {
-      const prevUnitSize = canvasConfigPrevRef.current?.unitSize;
-      setSnake(() => {
-        const newSnake = snakeRef.current?.map(({ x, y }) => ({
-          x: (x / prevUnitSize) * unitSize,
-          y: (y / prevUnitSize) * unitSize,
-        }));
-        return newSnake;
-      });
-
-      setFood(() => {
-        const newFood = {
-          x: Math.floor((foodRef.current.x / prevUnitSize) * unitSize),
-          y: Math.floor((foodRef.current.y / prevUnitSize) * unitSize),
-        };
-        return newFood;
-      });
-    }
-  }, []);
-
-  useEffect(() => {
-    setupCanvas();
-  }, [setupCanvas]);
-
-  useEffect(() => {
-    let resizeTimeout: ReturnType<typeof setTimeout> | null = null;
-
-    const handleResize = () => {
-      if (!resizeTimeout) {
-        resizeTimeout = setTimeout(() => {
-          setupCanvas();
-          resizeTimeout = null;
-        }, 200); // 200ms throttle delay
-      }
-    };
-
-    window.addEventListener("resize", handleResize);
-    return () => {
-      window.removeEventListener("resize", handleResize);
-      if (resizeTimeout) {
-        clearTimeout(resizeTimeout);
-      }
-    };
-  }, [setupCanvas]);
 
   const updateSnake = useCallback(() => {
     if (!canvasConfigRef.current) {
@@ -198,6 +93,20 @@ export default function SnakeGame({ win, gameOver, level, score, highScore }: Sn
       });
     }
   }, [snake, direction.x, direction.y, food.x, food.y, gameOver]);
+
+  //   const DIRECTION = {
+  //     LEFT: 'left',
+  //     TOP: 'top',
+  //     DOWN: 'down',
+  //     RIGHT: 'right',
+  //   } as const;
+  //
+  //   type Direction = typeof DIRECTION[keyof typeof DIRECTION];
+  // // 'left' | 'top' | 'down' | 'right'
+  //
+  // // Usage
+  //   const go = (dir: Direction) => console.log(`Going ${dir}`);
+  //   go(DIRECTION.TOP); // ✅
 
   const setDirectionFromEvents = useCallback((directionText: DirectionText) => {
     let directionTemp;
